refactor : refactor global css into component-wise css
