+++ src/components/Photo/photo.css
@@ -1,0 +1,49 @@
+.photo {
+  height: 17.5rem;
+  position: relative;
+  overflow: hidden;
+  cursor: pointer;
+}
+.photo > img {
+  width: 100%;
+  height: 100%;
+  display: block;
+  object-fit: cover;
+}
+.photo h4 {
+  margin-bottom: 0.5rem;
+}
+.photo p {
+  margin-bottom: 0;
+}
+
+.photo-info {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+.photo-info {
+  position: absolute;
+  width: 100%;
+  padding: 1rem;
+  bottom: 0;
+  left: 0;
+  background: rgba(0, 0, 0, 0.4);
+  color: white;
+  transform: translateY(100%);
+  transition: var(--transition);
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+.photo:hover .photo-info {
+  transform: translateY(0);
+}
+
+.bookmark {
+  color: var(--clr-white);
+}
+
+.bookmark:hover {
+  color: var(--clr-red-dark);
+}
