+++ src/components/Photos/photos.css
@@ -1,0 +1,20 @@
+.photos {
+  padding: 5rem 0;
+  margin-top: 1.5rem;
+}
+.photos-center {
+  width: 90vw;
+  max-width: var(--max-width);
+  margin: 0 auto;
+  display: grid;
+  gap: 2rem;
+}
+
+@media screen and (min-width: 576px) {
+  .photos-center {
+    grid-template-columns: repeat(auto-fill, minmax(368px, 1fr));
+  }
+  .search-form {
+    max-width: var(--fixed-width);
+  }
+}
