+++ src/components/ProtectedRoute/ProtectedRoute.js
@@ -2,7 +2,7 @@ import { useAuth0 } from "@auth0/auth0-react";
 import React from "react";
 import { Route, Redirect } from "react-router-dom";
 import { useSnackbar } from "react-simple-snackbar";
-import { options } from "../utils/options";
+import { options } from "../../utils/options";
 
 const ProtectedRoute = ({ children, ...rest }) => {
   const { isAuthenticated } = useAuth0();
