+++ src/components/SavedPhoto/savedphoto.css
@@ -1,0 +1,91 @@
+.saved-photo {
+  height: 31rem;
+  background: var(--clr-grey-10);
+  transition: var(--transition);
+  border-radius: var(--radius);
+}
+
+.saved-photo:hover {
+  box-shadow: var(--dark-shadow);
+  transform: scale(1.025);
+}
+
+.saved-photo > img {
+  width: 100%;
+  height: 24rem;
+  display: block;
+  object-fit: cover;
+}
+
+.saved-photo-header {
+  display: flex;
+  justify-content: space-between;
+  padding: 0.5rem 0.5rem;
+}
+
+.user-img {
+  width: 2.5rem;
+  height: 2.5rem;
+  border-radius: 50%;
+  margin-right: 0.5rem;
+}
+
+.user-info {
+  display: flex;
+  align-items: center;
+  color: var(--clr-primary-3);
+}
+.user-info > p {
+  margin-bottom: 0;
+}
+
+.close-btn-2 {
+  background: transparent;
+  border: none;
+  color: red;
+  font-size: 2rem;
+  cursor: pointer;
+}
+
+.saved-photo-footer {
+  display: flex;
+  justify-content: space-between;
+  padding: 0.5rem 0.5rem;
+  border-radius: var(--radius);
+}
+
+.link-btn {
+  border: transparent;
+  color: var(--clr-grey-1);
+  background: var(--clr-primary-9);
+  border-radius: var(--radius);
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  font-size: 1rem;
+  padding: 0.25rem;
+  cursor: pointer;
+}
+
+.link-btn-icon {
+  display: flex;
+  align-items: center;
+  margin-right: 0.4rem;
+}
+
+.empty-images {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-top: 5rem;
+}
+
+.empty-images > h3 {
+  color: var(--clr-primary-5);
+}
+
+.empty-img {
+  display: block;
+  width: 50vw;
+  height: 100vh;
+}
