+++ src/components/SearchForm/searchform.css
@@ -1,0 +1,38 @@
+.search {
+  padding: 5rem 0 0 0;
+  width: 90vw;
+  max-width: var(--max-width);
+  margin: 0 auto;
+  margin-top: 3rem;
+}
+.search-form {
+  display: flex;
+}
+.form-input {
+  width: 100%;
+  outline: transparent;
+}
+.form-input,
+.submit-btn {
+  padding: 0.75rem 1.25rem;
+  border: none;
+  text-transform: capitalize;
+  letter-spacing: var(--spacing);
+  font-size: 1.5rem;
+  background: transparent;
+  color: var(--clr-grey-5);
+  border-bottom: 3px solid var(--clr-grey-5);
+  cursor: pointer;
+}
+.form-input {
+  color: var(--clr-grey-3);
+}
+.form-input::placeholder {
+  color: var(--clr-grey-5);
+}
+
+@media screen and (min-width: 576px) {
+  .search-form {
+    max-width: var(--fixed-width);
+  }
+}
