+++ src/index.css
@@ -135,256 +135,9 @@ p {
 
 /*
 =============== 
-Stock Photos
+Saved Photos
 ===============
 */
-
-.navbar {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  background: var(--clr-white);
-  box-shadow: var(--dark-shadow);
-  z-index: 2;
-}
-
-.nav-center {
-  width: 100%;
-  max-width: 1170px;
-  margin: 0 auto;
-}
-
-.nav-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 1rem;
-}
-
-.nav-toggle {
-  background: transparent;
-  border: transparent;
-  font-size: 1.75rem;
-  color: var(--clr-primary-5);
-  cursor: pointer;
-  transition: var(--transition);
-}
-
-.nav-toggle:hover {
-  transform: scale(1.2);
-}
-
-.nav-links {
-  height: 0;
-  overflow: hidden;
-  transition: var(--transition);
-}
-
-.nav-footer {
-  /* padding: 1rem 2rem; */
-  display: none;
-}
-.show {
-  display: block;
-}
-
-.show-links {
-  height: 200px;
-}
-.nav-link {
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-  border: transparent;
-  background: transparent;
-  padding: 1rem 2rem;
-  text-transform: capitalize;
-  letter-spacing: var(--spacing);
-  transition: var(--transition);
-  color: var(--clr-grey-1);
-  font-size: 1rem;
-  cursor: pointer;
-}
-
-.nav-btn {
-  width: 100%;
-}
-
-.generic-icon {
-  background: transparent;
-  border: transparent;
-  font-size: 1.5rem;
-  margin-right: 0.5rem;
-  cursor: pointer;
-}
-
-.logo {
-  font-size: 1.75rem;
-  color: var(--clr-primary-5);
-}
-
-.nav-link:hover {
-  color: var(--clr-primary-1);
-  background: var(--clr-primary-8);
-  padding-left: 2.25rem;
-}
-
-@media screen and (min-width: 900px) {
-  .nav-center {
-    display: flex;
-    justify-content: space-between;
-  }
-  .nav-toggle {
-    display: none;
-  }
-  .nav-links {
-    height: auto;
-    display: flex;
-    justify-content: space-between;
-    padding: 1rem 2rem;
-  }
-  .show-links {
-    height: auto;
-  }
-
-  .nav-link {
-    padding: 0;
-    margin-right: 1.5rem;
-  }
-
-  .nav-link:hover {
-    padding-left: 0rem;
-    color: var(--clr-primary-5);
-    background: transparent;
-  }
-
-  .nav-footer {
-    padding: 1rem 2rem;
-    display: block;
-  }
-}
-
-/*
-=============== 
-Search
-===============
-*/
-
-.search {
-  padding: 5rem 0 0 0;
-  width: 90vw;
-  max-width: var(--max-width);
-  margin: 0 auto;
-  margin-top: 3rem;
-}
-.search-form {
-  display: flex;
-}
-.form-input {
-  width: 100%;
-  outline: transparent;
-}
-.form-input,
-.submit-btn {
-  padding: 0.75rem 1.25rem;
-  border: none;
-  text-transform: capitalize;
-  letter-spacing: var(--spacing);
-  font-size: 1.5rem;
-  background: transparent;
-  color: var(--clr-grey-5);
-  border-bottom: 3px solid var(--clr-grey-5);
-  cursor: pointer;
-}
-.form-input {
-  color: var(--clr-grey-3);
-}
-.form-input::placeholder {
-  color: var(--clr-grey-5);
-}
-
-/*
-=============== 
-Photos
-===============
-*/
-
-.photos {
-  padding: 5rem 0;
-  margin-top: 1.5rem;
-}
-.photos-center {
-  width: 90vw;
-  max-width: var(--max-width);
-  margin: 0 auto;
-  display: grid;
-  gap: 2rem;
-}
-.photo {
-  height: 17.5rem;
-  position: relative;
-  overflow: hidden;
-  cursor: pointer;
-}
-.photo > img {
-  width: 100%;
-  height: 100%;
-  display: block;
-  object-fit: cover;
-}
-.photo h4 {
-  margin-bottom: 0.5rem;
-}
-.photo p {
-  margin-bottom: 0;
-}
-
-.photo-info {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-.photo-info {
-  position: absolute;
-  width: 100%;
-  padding: 1rem;
-  bottom: 0;
-  left: 0;
-  background: rgba(0, 0, 0, 0.4);
-  color: white;
-  transform: translateY(100%);
-  transition: var(--transition);
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-.photo:hover .photo-info {
-  transform: translateY(0);
-}
-
-.bookmark {
-  color: var(--clr-white);
-}
-
-.bookmark:hover {
-  color: var(--clr-red-dark);
-}
-@media screen and (min-width: 576px) {
-  .photos-center {
-    grid-template-columns: repeat(auto-fill, minmax(368px, 1fr));
-  }
-  .search-form {
-    max-width: var(--fixed-width);
-  }
-}
-
-/*
-=============== 
-Saved Photo
-===============
-*/
-
 .saved-photos-center {
   width: 90vw;
   max-width: var(--max-width);
@@ -393,176 +146,10 @@ Saved Photo
   column-gap: 2rem;
   row-gap: 2rem;
 }
-
-.saved-photo {
-  height: 31rem;
-  background: var(--clr-grey-10);
-  transition: var(--transition);
-  border-radius: var(--radius);
-}
-
-.saved-photo:hover {
-  box-shadow: var(--dark-shadow);
-  transform: scale(1.025);
-}
-
-.saved-photo > img {
-  width: 100%;
-  height: 24rem;
-  display: block;
-  object-fit: cover;
-}
-
-.saved-photo-header {
-  display: flex;
-  justify-content: space-between;
-  padding: 0.5rem 0.5rem;
-}
-
-.user-img {
-  width: 2.5rem;
-  height: 2.5rem;
-  border-radius: 50%;
-  margin-right: 0.5rem;
-}
-
-.user-info {
-  display: flex;
-  align-items: center;
-  color: var(--clr-primary-3);
-}
-.user-info > p {
-  margin-bottom: 0;
-}
-
-.close-btn-2 {
-  background: transparent;
-  border: none;
-  color: red;
-  font-size: 2rem;
-  cursor: pointer;
-}
-
-.saved-photo-footer {
-  display: flex;
-  justify-content: space-between;
-  padding: 0.5rem 0.5rem;
-  border-radius: var(--radius);
-}
-
-.link-btn {
-  border: transparent;
-  color: var(--clr-grey-1);
-  background: var(--clr-primary-9);
-  border-radius: var(--radius);
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  font-size: 1rem;
-  padding: 0.25rem;
-  cursor: pointer;
-}
-
-.link-btn-icon {
-  display: flex;
-  align-items: center;
-  margin-right: 0.4rem;
-}
-
-.empty-images {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-top: 5rem;
-}
-
-.empty-images > h3 {
-  color: var(--clr-primary-5);
-}
-
-.empty-img {
-  display: block;
-  width: 50vw;
-  height: 100vh;
-}
-
 @media screen and (min-width: 576px) {
   .saved-photos-center {
     grid-template-columns: repeat(auto-fill, minmax(368px, 1fr));
   }
-}
-/*
-=============== 
-Image Viewer
-===============
-*/
-.image-viewer-overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  height: 100%;
-  width: 100%;
-  background: rgba(0, 0, 0, 0.8);
-  display: grid;
-  place-items: center;
-  z-index: -5;
-  visibility: hidden;
-}
-
-.show-image-viewer {
-  z-index: 5;
-  visibility: visible;
-}
-
-.image-display-container {
-  display: block;
-  height: 80vh;
-  width: 80vw;
-  /* background: var(--clr-white); */
-  background: transparent;
-}
-
-.image-header {
-  display: flex;
-  padding: 0 2rem;
-  align-items: center;
-}
-.close-btn {
-  background: transparent;
-  border: none;
-  color: red;
-  font-size: 3rem;
-  cursor: pointer;
-  margin-left: auto;
-}
-
-.image-display {
-  display: flex;
-  justify-content: space-between;
-  height: 90%;
-  width: 100%;
-}
-
-.image-btns-container {
-  align-self: center;
-}
-
-.image-btns {
-  background: transparent;
-  border: none;
-  color: var(--clr-primary-4);
-  font-size: 3rem;
-  cursor: pointer;
-}
-
-.image-box {
-  background-color: transparent;
-}
-
-.current-image {
-  height: 100%;
-  width: 100%;
-  object-fit: cover;
 }
 
 /*
@@ -588,27 +175,4 @@ Error
   color: var(--clr-grey-2);
   background: var(--clr-primary-7);
   border-radius: var(--radius);
-}
-
-/*
-=============== 
-Loading
-===============
-*/
-
-@keyframes spinner {
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-.loading {
-  width: 6rem;
-  height: 6rem;
-  margin: 0 auto;
-  margin-top: 10rem;
-  border-radius: 50%;
-  border: 4px solid #ccc;
-  border-top-color: var(--clr-primary-5);
-  animation: spinner 0.6s linear infinite;
 }
