+++ src/pages/Authwrapper.js
@@ -1,6 +1,6 @@
 import { useAuth0 } from "@auth0/auth0-react";
 import React from "react";
-import Loading from "../components/Loading";
+import Loading from "../components/Loading/Loading";
 
 const Authwrapper = ({ children }) => {
   const { isLoading, error } = useAuth0();
